refactor(sfc): split inline compiler phases
